backend-python/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
sqlalchemy==2.0.20
alembic==1.12.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.3.0
